+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add konf.WithCaseSensitive to support case-sensitive path match (#205).
+
 ## [0.8.1] - 2024-03-06
 
 ### Fixed